third_party/blink/web_tests/platform/generic/fast/css/invalidation/targeted-class-type-selectors-expected.txt

Check that targeted class invalidation works for type selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(children[0][0], null).backgroundColor is transparent
PASS getComputedStyle(children[0][1], null).backgroundColor is transparent
PASS getComputedStyle(children[0][2], null).backgroundColor is transparent
PASS getComputedStyle(children[0][3], null).backgroundColor is transparent
PASS getComputedStyle(children[1][0], null).backgroundColor is transparent
PASS getComputedStyle(children[1][1], null).backgroundColor is transparent
PASS getComputedStyle(children[1][2], null).backgroundColor is transparent
PASS getComputedStyle(children[1][3], null).backgroundColor is transparent
PASS getComputedStyle(children[2][0], null).backgroundColor is transparent
PASS getComputedStyle(children[2][1], null).backgroundColor is transparent
PASS getComputedStyle(children[2][2], null).backgroundColor is transparent
PASS getComputedStyle(children[2][3], null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(children[0][0], null).backgroundColor is green
PASS getComputedStyle(children[0][1], null).backgroundColor is green
PASS getComputedStyle(children[0][2], null).backgroundColor is transparent
PASS getComputedStyle(children[0][3], null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(children[1][0], null).backgroundColor is transparent
PASS getComputedStyle(children[1][1], null).backgroundColor is transparent
PASS getComputedStyle(children[1][2], null).backgroundColor is green
PASS getComputedStyle(children[1][3], null).backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 5
PASS getComputedStyle(children[2][0], null).backgroundColor is green
PASS getComputedStyle(children[2][1], null).backgroundColor is green
PASS getComputedStyle(children[2][2], null).backgroundColor is green
PASS getComputedStyle(children[2][3], null).backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
